LAB1/requirements.txt
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
pillow==10.0.1
matplotlib==3.7.2
requests==2.31.0
python-multipart==0.0.6
seaborn==0.13.2